CacheMethod is back
Went back to the caching formulation again to make it simpler to call
things. Though I'm not sure this is good because there are so many
problems with how the cache works (it's confusing).

This is a trade-off between calculating the hashes all the time and
calculating an extra gradient, hessian and predicted on most algorithms.